+++ cli/CHANGELOG.md
@@ -4,16 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.4.0 (2021-07-12)
+### Changed
+- Switch to renamed `yubikey` crate ([#283])
+- Bump MSRV to 1.51+ ([#283])
+
+[#283]: https://github.com/iqlusioninc/yubikey.rs/pull/283
+
 ## 0.3.0 (2021-03-22)
 ### Changed
-- Bump `yubikey` dependency to v0.3 ([#240])
+- Bump `yubikey-piv` dependency to v0.3 ([#240])
 
 [#240]: https://github.com/iqlusioninc/yubikey.rs/pull/240
 
 ## 0.2.0 (2021-01-30)
 ### Changed
 - Bump MSRV to 1.46+ ([#208])
-- Bump `yubikey` dependency to v0.2 ([#220])
+- Bump `yubikey-piv` dependency to v0.2 ([#220])
 
 [#208]: https://github.com/iqlusioninc/yubikey.rs/pull/208
 [#220]: https://github.com/iqlusioninc/yubikey.rs/pull/220
@@ -23,7 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `status` command ([#72], [#74])
 
 ### Changed
-- Bump `yubikey` to v0.1.0 ([#180])
+- Bump `yubikey-piv` to v0.1.0 ([#180])
 - Bump `x509-parser` to v0.8 ([#181])
 - Bump `sha2` to v0.9 ([#182])
 - Rename `list` command to `readers`; improve usage ([#71])
